Fix pending doc-update jobs being canceled
There is a stupid GitHub feature that cancels pending jobs in a
concurrency group... even if you don't ask for them to be canceled.
See: https://github.com/orgs/community/discussions/5435

It seems I'm not the only one puzzled by this.

Anyway, this messes up for example dataset updates, as it causes them
to be dropped completely. This commit disables the concurrency groups
entirely, which should still be fine, as we use rebasing in the CI.

Issue: https://github.com/RiverBench/RiverBench/issues/67